fix: Fix the user role assignment
Previously the role assignment was managed inside the domain entity manually. We moved this logic to a dedicated service with proper Dependency Injection, that is now responsible to deal with roles management.
We should mention that there are no automatic role assigment, and a call to the IAuthManager service is mendatory in order to manage the roles of a specific user