fix: InMemory Cache setup for user existance Middleware
There is a middleware that checks for each authenticated request if the attached user still exists in the database.
This middleware to work properly needed to make a database call for each request, that would be performance inefficient. The solution was to introduce a cache system that will store the state of a user and retrieve it for each subsequent request